expose HTTP endpoint to trigger a poll
