+++ cmd/server/main.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"os"
+
+	"github.com/utkuufuk/entrello/internal/config"
+	"github.com/utkuufuk/entrello/internal/logger"
+	"github.com/utkuufuk/entrello/internal/service"
+)
+
+func main() {
+	http.HandleFunc("/", handler)
+	http.ListenAndServe(fmt.Sprintf(":%s", os.Getenv("PORT")), nil)
+}
+
+func handler(w http.ResponseWriter, req *http.Request) {
+	if req.Method != http.MethodPost {
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		return
+	}
+
+	body, err := ioutil.ReadAll(req.Body)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	var cfg config.Config
+	if err = json.Unmarshal(body, &cfg); err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	if err = service.Poll(cfg); err != nil {
+		logger.Error(err.Error())
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	w.WriteHeader(http.StatusOK)
+}
