+++ internal/config/config.go
@@ -42,17 +42,6 @@ type Config struct {
 }
 
 func ReadConfig(fileName string) (cfg Config, err error) {
-	jsonStr := os.Getenv("ENTRELLO_CONFIG")
-	if jsonStr != "" {
-		logger.Info("Attempting to read configuration from environment variable 'ENTRELLO_CONFIG'")
-		err = json.Unmarshal([]byte(jsonStr), &cfg)
-		if err != nil {
-			return cfg, fmt.Errorf("could not parse config stored in the environment variable: %s", err)
-		}
-
-		return cfg, nil
-	}
-
 	logger.Info("Attempting to read configuration from file '%s'", fileName)
 	f, err := os.Open(fileName)
 	if err != nil {
