+++ internal/service/source.go
@@ -1,4 +1,4 @@
-package main
+package service
 
 import (
 	"encoding/json"
